backend/requirements.txt
fastapi
uvicorn
redis
psycopg2-binary
python-dotenv
openfoodfacts
python-multipart
pydantic
azure-cosmos
azure-identity
behave
transformers
torch
requests